+++ app/Http/Controllers/Admin/LanguageController.php
@@ -7,12 +7,19 @@ use Illuminate\Http\Request;
 use App\Http\Requests\Admin\StoreSkillRequest;
 use App\Http\Requests\Admin\LanguageRequest;
 use App\Models\Language;
+use App\Models\JobPostLanguage;
+use App\Models\SeekerLanguage;
 
 class LanguageController extends Controller
 {
     /**
      * Display a listing of the resource.
      */
+    public function __construct()
+    {
+        // Áp dụng middleware 'custom' cho các phương thức 'update' và 'destroy'
+        $this->middleware(['checkAdminType'])->only(['update', 'destroy']);
+    }
     public function index()
     {
         $major = Language::paginate(10);
@@ -87,6 +94,13 @@ class LanguageController extends Controller
         return redirect()->route('admin.language.index')->with('success', 'Khôi phục thành công');
     }
     public function force(string $id){
+        if(auth('admin')->user()->type==0)
+            return redirect()->route('admin.language.trash')->with('error', 'Bạn không có quyền xóa dữ liệu');
+        //đếm xem có bảng nào tham chiếu đến chuyên ngành này không
+        $count_seeker_language = SeekerLanguage::where('language_id',$id)->get()->count();
+        $count_job_post_language = JobPostLanguage::where('language_id',$id)->get()->count();
+        if($count_seeker_language || $count_job_post_language)
+            return redirect()->route('admin.language.trash')->with('error', 'Không thể xóa, sẽ lỗi trang web');
         Language::withTrashed()->where('id', $id)->forceDelete();
         return redirect()->route('admin.language.trash')->with('success', 'Xóa thành công');
