+++ app/Http/Controllers/Admin/TimeController.php
@@ -4,8 +4,10 @@ namespace App\Http\Controllers\Admin;
 
 use App\Http\Controllers\Controller;
 use App\Models\TimeExperience;
+use App\Models\JobPost;
 use Illuminate\Http\Request;
 use App\Http\Requests\Admin\DegreeRequest;
+use App\Http\Requests\Admin\TimeRequest;
 
 
 class TimeController extends Controller
@@ -13,6 +15,11 @@ class TimeController extends Controller
     /**
      * Display a listing of the resource.
      */
+    public function __construct()
+    {
+        // Áp dụng middleware 'custom' cho các phương thức 'update' và 'destroy'
+        $this->middleware(['checkAdminType'])->only(['update', 'destroy']);
+    }
     public function index()
     {
         $title = 'Kinh nghiệm';
@@ -37,7 +44,7 @@ class TimeController extends Controller
     /**
      * Store a newly created resource in storage.
      */
-    public function store(DegreeRequest $request)
+    public function store(TimeRequest $request)
     {
         try {
             TimeExperience::create($request->all());
@@ -67,7 +74,7 @@ class TimeController extends Controller
     /**
      * Update the specified resource in storage.
      */
-    public function update(DegreeRequest $request, TimeExperience $time)
+    public function update(TimeRequest $request, TimeExperience $time)
     {
         try {
             $time->update($request->all());
@@ -103,6 +110,13 @@ class TimeController extends Controller
         return redirect()->route('admin.time.index')->with('success', 'Khôi phục thành công');
     }
     public function force(string $id){
+        if(auth('admin')->user()->type==0)
+            return redirect()->route('admin.time.trash')->with('error', 'Bạn không có quyền xóa dữ liệu');
+        //đếm xem có bảng nào tham chiếu đến chuyên ngành này không
+        // $count_edu = Education::where('major_id',$id)->get()->count();
+        $count_job_post = JobPost::where('time_exp_id',$id)->get()->count();
+        if($count_job_post)
+            return redirect()->route('admin.time.trash')->with('error', 'Không thể xóa, sẽ lỗi trang web');
         TimeExperience::withTrashed()->where('id', $id)->forceDelete();
         // return response()->json(['success'=>'Xóa thành công!']);
         return redirect()->route('admin.time.trash')->with('success', 'Xóa thành công');
